Update index.html
Rewrite for the page so that it has tabs and a beautiful background thanks to Elliott Engelmann from Unsplash (https://unsplash.com/photos/silhouette-of-mountain-DjlKxYFJlTc).
Also thanks to DeepSeek for tons of iterations to get this just right.

No version numbers for now since this is far from done and not even sure if and how the Google Form will work under this webpage. Oh and you can now go straight to imaginowsky.github.io to see if it works.