Update index.html
Yet another update. This time the layout has been a bit improved, submit button has been redone and instead of going to the google forms confirmation page, it confirms the form being submitted with a cool prompt and clears the form. The form works to be filled again and submitted without reloading the page!